requirements-mcp.txt
# Minimal requirements for MCP Server and FastAPI Backend
# Use this for a clean, working installation

# Core FastAPI
fastapi==0.121.2
uvicorn==0.35.0
pydantic==2.11.7
pydantic-settings>=2.0.0

# MCP Server
mcp==1.21.0
httpx==0.28.1
python-dotenv==1.1.1

# AI Services
google-generativeai==0.8.5
anthropic==0.64.0

# Database & Cache
neo4j==5.28.2
redis==6.4.0

# Logging & Monitoring
structlog==25.4.0
prometheus-client>=0.17.0

# Additional API dependencies
starlette>=0.40.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Async support
anyio>=4.0.0
